+++ src/main/java/capstone/voicereport/repository/VoiceReportRepository.java
@@ -8,9 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface VoiceReportRepository extends JpaRepository<VoiceReport, Integer> {
-
-
+public interface VoiceReportRepository extends JpaRepository<VoiceReport, Long> {
 
     @Query(
             value = """
